navidades aimar 04/01/2024

+++ app/Http/Controllers/CycleController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Models\Cycle;
+use App\Models\Module;
+use App\Models\Department;
 use Illuminate\Http\Request;
 
 class CycleController extends Controller
@@ -10,12 +12,30 @@ class CycleController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
+
+        if ($request->is('admin*')) {
+            $perPage = $request->input('per_page', 10);
+            $cycles = Cycle::orderBy('name', 'asc')->paginate($perPage);
+            foreach ($cycles as $cycle) {
+                $cycle->count_students = Cycle::join('cycle_users', 'cycles.id', '=', 'cycle_users.cycle_id')
+                    ->join('users', 'cycle_users.user_id', '=', 'users.id')
+                    ->join('role_users', 'users.id', '=', 'role_users.user_id')
+                    ->where('role_users.role_id', '=', 3)
+                    ->where('cycles.id', '=', $cycle->id)
+                    ->count();
+                $cycle->department = Department::find($cycle->department_id);
+            }
+            $cycles->totalCycles = Cycle::count();
+            return view('admin.cycles.index', compact('cycles'));
+        }else {
+            //si no es admin
+
+        }
         $cycle = Cycle::orderBy('name','asc')->get();
         return view('cycles.index',['cycles'=>$cycle]);
     }
-
     /**
      * Show the form for creating a new resource.
      */
@@ -38,8 +58,46 @@ class CycleController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Cycle $cycle)
+    public function show(Request $request, Cycle $cycle)
     {
+        if ($request->is('admin*')) {
+            $perPage = $request->input('per_page', 10);
+            $cycleId = $cycle->id;
+
+            $cycle->count_students = Cycle::join('cycle_users', 'cycles.id', '=', 'cycle_users.cycle_id')
+                ->join('users', 'cycle_users.user_id', '=', 'users.id')
+                ->join('role_users', 'users.id', '=', 'role_users.user_id')
+                ->where('role_users.role_id', '=', 3)
+                ->where('cycles.id', '=', $cycleId)
+                ->count();
+
+            $cycle->modules = Module::join('cycle_module', 'modules.id', '=', 'cycle_module.module_id')
+                ->where('cycle_module.cycle_id', $cycleId)
+                ->select('modules.*')
+                ->paginate($perPage);
+
+            foreach ($cycle->modules as $module) {
+                $module->count_teachers = Module::join('module_user', 'modules.id', '=', 'module_user.module_id')
+                    ->join('users', 'module_user.user_id', '=', 'users.id')
+                    ->join('role_users', 'users.id', '=', 'role_users.user_id')
+                    ->where('role_users.role_id', '=', 2)
+                    ->where('modules.id', '=', $module->id)
+                    ->count();
+                $module->count_students = Module::join('module_user', 'modules.id', '=', 'module_user.module_id')
+                    ->join('users', 'module_user.user_id', '=', 'users.id')
+                    ->join('role_users', 'users.id', '=', 'role_users.user_id')
+                    ->where('role_users.role_id', '=', 3)
+                    ->where('modules.id', '=', $module->id)
+                    ->count();
+
+            }
+            $cycle->department = Department::find($cycle->department_id);
+            return view('admin.cycles.show', compact('cycle'));
+        }else {
+            //si no es admin
+
+        }
+
         return view('cycles.show',['cycle'=>$cycle]);
     }
 
@@ -65,9 +123,20 @@ class CycleController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Cycle $cycle)
+    public function destroy(Request $request, $cycleId)
     {
-        $cycle->delete();
-        return redirect()->route('cycles.index');
+        $redirectRoute = 'cycles.index';
+
+        if ($request->is('admin*')) {
+            //si es admin
+            $cycle = Cycle::find($cycleId);
+            if ($cycle) {
+                $cycle->delete();
+                return redirect()->route($redirectRoute)->with('success', 'Usuario eliminado exitosamente.');
+            } else {
+                return redirect()->back()->with('error', 'Usuario no encontrado.');
+            }
+        }else {
+        }
     }
 }
