+++ app/Http/Controllers/RoleController.php
@@ -3,18 +3,30 @@
 namespace App\Http\Controllers;
 
 use App\Models\Role;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Mail\Message;
+use Illuminate\Support\Facades\DB;
 
 class RoleController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        $roles = Role::orderBy('name','asc')->get();
-        return view('roles.index',['roles'=>$roles]);
+
+        if ($request->is('admin*')) {
+            $perPage = $request->input('per_page', 10);
+            $roles = Role::orderBy('name', 'asc')->paginate($perPage);
+            foreach ($roles as $role) {
+                $role->count_people = DB::table('role_users')->where('role_id', $role->id)->count();
+            }
+            return view('admin.roles.index', compact('roles'));
+        }else {
+            //si no es admin
+
+        }
     }
 
     /**
@@ -22,7 +34,7 @@ class RoleController extends Controller
      */
     public function create()
     {
-        return view('roles.create');
+        return view('admin.roles.create');
     }
 
     /**
@@ -39,9 +51,15 @@ class RoleController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Role $role)
+    public function show(Request $request, Role $role)
     {
-        return view('roles.show',['role'=>$role]);
+        $perPage = $request->input('per_page', 10);
+        $users = User::whereHas('roles', function ($query) use ($role) {
+            $query->where('id', $role->id);
+        })->paginate($perPage);
+
+        $role->users = $users;
+        return view('admin.roles.show',['role'=>$role]);
     }
 
     /**
@@ -66,13 +84,23 @@ class RoleController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Role $role)
+    public function destroy(Request $request, $roleId)
     {
-        if($role->name != 'ADMINISTRADOR' && $role->name!= 'ALUMNO' && $role->name!= 'PROFESOR'){
-            $role->delete();
-            return redirect()->route('roles.index');
-        }else{
-            return ('No se puede eliminar el rol ADMINISTRADOR o ALUMNO o PROFES');
+        $redirectRoute = 'roles.index';
+
+        if ($request->is('admin*')) {
+            $role = Role::find($roleId);
+            if ($role) {
+                if($role->name != 'ADMINISTRADOR' && $role->name!= 'ALUMNO' && $role->name!= 'PROFESOR'){
+                    $role->delete();
+                    return redirect()->route($redirectRoute)->with('success', 'Rol eliminado exitosamente.');
+                }else{
+                    return redirect()->back()->with('error', 'No se puede eliminar el rol ADMINISTRADOR o ALUMNO o PROFES');
+                }
+            } else {
+                return redirect()->back()->with('error', 'Departamento no encontrado.');
+            }
+        }else {
         }
     }
 }
